// src/types/mealPreferences.ts
export type MealPreferences = {
  cookingTimePerMeal?: string;
  breakfastTime?: string;
  lunchTime?: string;
  dinnerTime?: string;
  snackPreference?: string;
  favoriteProteins?: string[];
  favoriteCarbs?: string[];
  favoriteVeggies?: string[];
  favoriteFruits?: string[];
  allergies?: string;
  additionalNotes?: string;
};